add and update teacher

+++ src/components/Table/TaskTable.jsx
@@ -5,7 +5,6 @@ import {
   flexRender,
   getCoreRowModel,
   getFilteredRowModel,
-  getPaginationRowModel,
   getSortedRowModel,
   useReactTable,
 } from "@tanstack/react-table";
@@ -13,6 +12,8 @@ import EditableCell from "./EditableCell";
 import StatusCell from "./StatusCell";
 import Filters from "./Filters";
 import SortIcon from "./icons/SortIcon";
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 const columns = [
   {
@@ -73,14 +74,14 @@ const TaskTable = ({ items }) => {
   }, [search])
 
   useEffect(() => {
-    if (!isMounted && items.length > 0) { 
+    if (!isMounted && items.length > 0) {
       setData(extractContent(items));
-      setIsMounted(true); 
+      setIsMounted(true);
     } else {
       fetchData();
     }
-  }, [items, isMounted, pageIndex]);
-  
+  }, [pageIndex]);
+
   const fetchData = (name = "") => {
     try {
       store.getTeacherData(pageIndex, name)
@@ -111,17 +112,71 @@ const TaskTable = ({ items }) => {
     return (Array.isArray(array) && array.length > 0 && Array.isArray(array[0].content)) ? array[0].content : []
   }
 
-  const deleteRow = (rowIndex, id) => {
-    store.deleteById(id).then(() => {
+  const deleteRow = (rowIndex, id, isAdd) => {
+    if(isAdd) {
       const newData = [...data];
       newData.splice(rowIndex, 1);
       setData(newData);
-    }).catch(e => console.log(e.data))
+    }else {
+      store.deleteById(id).then(() => {
+        const newData = [...data];
+        newData.splice(rowIndex, 1);
+        setData(newData);
+        toast.success("Success")
+      }).catch(err => {
+        toast.error("Failed")
+      })
+    }
   };
+
+  const saveRow = (rowIndex, row) => {
+    if (!row.firstName && !row.lastName && !row.departments.length) {
+      toast.warning("Validation error")
+      return
+    }
+
+    const newRow = {
+      firstName: row.firstName,
+      lastName: row.lastName,
+      middleName: row.middleName?.length ? row.middleName : undefined,
+      cabinet: row.cabinet?.length ? row.cabinet : undefined,
+      departmentIds: row.departments.map(dep => dep.id),
+    };
+
+    console.log(newRow)
+
+    if(row.isAdd === true) {
+      store.add(newRow).then(() => {
+        const newData = [...data];
+        row["isAdd"] = false
+        newData[rowIndex] = row;
+        setData(newData);
+
+        toast.success("Success")
+      }).catch((err) => {
+        console.log(err)
+        toast.error("Failed")
+      })
+    }else if(row.isUpdate === true) {
+      store.update(row.id, newRow).then((d) => {
+        console.log(d)
+        const newData = [...data];
+        row["isUpdate"] = false
+        newData[rowIndex] = row;
+        setData(newData);
+
+        toast.success("Success")
+      }).catch((err) => {
+        console.log(err)
+        toast.error("Failed")
+      })
+    }
+  }
 
   const addRow = () => {
     setData([...data, {
-      departments: []
+      departments: [],
+      isAdd: true
     }]);
   };
 
@@ -140,6 +195,7 @@ const TaskTable = ({ items }) => {
               ? {
                   ...prev[rowIndex],
                   [columnId]: value,
+                  isUpdate: !row.isAdd
                 }
               : row
           )
@@ -148,85 +204,90 @@ const TaskTable = ({ items }) => {
   });
 
   return (
-    <Box>
-      <Filters
-        value={search}
-        onValueChange={setSearch}
-      />
-      <Box className="table" w={table.getTotalSize()}>
-        {table.getHeaderGroups().map((headerGroup) => (
-          <Box className="tr" key={headerGroup.id}>
-            {headerGroup.headers.map((header) => (
-              <Box className="th" w={header.getSize()} key={header.id}>
-                {header.column.columnDef.header}
-                {header.column.getCanSort() && (
-                  <Icon
-                    as={SortIcon}
-                    mx={3}
-                    fontSize={14}
-                    onClick={header.column.getToggleSortingHandler()}
-                  />
-                )}
-                {
-                  {
-                    asc: " 🔼",
-                    desc: " 🔽",
-                  }[header.column.getIsSorted()]
+      <Box>
+        <Filters
+            value={search}
+            onValueChange={setSearch}
+        />
+        <Box className="table" w={table.getTotalSize()}>
+          {table.getHeaderGroups().map((headerGroup) => (
+              <Box className="tr" key={headerGroup.id}>
+                {headerGroup.headers.map((header) => (
+                    <Box className="th" w={header.getSize()} key={header.id}>
+                      {header.column.columnDef.header}
+                      {header.column.getCanSort() && (
+                          <Icon
+                              as={SortIcon}
+                              mx={3}
+                              fontSize={14}
+                              onClick={header.column.getToggleSortingHandler()}
+                          />
+                      )}
+                      {
+                        {
+                          asc: " 🔼",
+                          desc: " 🔽",
+                        }[header.column.getIsSorted()]
+                      }
+                      <Box
+                          onMouseDown={header.getResizeHandler()}
+                          onTouchStart={header.getResizeHandler()}
+                          className={`resizer ${
+                              header.column.getIsResizing() ? "isResizing" : ""
+                          }`}
+                      />
+                    </Box>
+                ))}
+              </Box>
+          ))}
+          {table.getRowModel().rows.map((row, rowIndex) => (
+              <Box className="tr" key={row.id}>
+                {row.getVisibleCells().map((cell) => (
+                    <Box className="td" w={cell.column.getSize()} key={cell.id}>
+                      {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                    </Box>
+                ))}
+                {(row.original.isAdd) &&
+                    <Box w="100px" padding="5px">
+                      <Button onClick={() => {
+                        deleteRow(rowIndex, null, true)
+                      }}>Удалить</Button>
+                    </Box>
                 }
-                <Box
-                  onMouseDown={header.getResizeHandler()}
-                  onTouchStart={header.getResizeHandler()}
-                  className={`resizer ${
-                    header.column.getIsResizing() ? "isResizing" : ""
-                  }`}
-                />
+
+                {(row.original.isUpdate || row.original.isAdd) &&
+                    <Box w="100px" padding="5px">
+
+                      <Button onClick={() => {
+                        saveRow(rowIndex, row.original)
+                      }}>Сохранить</Button>
+
+                    </Box>
+                }
               </Box>
-            ))}
-          </Box>
-        ))}
-        {table.getRowModel().rows.map((row, rowIndex) => (
-          <Box className="tr" key={row.id}>
-            {row.getVisibleCells().map((cell) => (
-              <Box className="td" w={cell.column.getSize()} key={cell.id}>
-                {flexRender(cell.column.columnDef.cell, cell.getContext())}
-              </Box>
-            ))}
-            <Box w="100px">
-              <Box
-                pos="absolute"
-                cursor="pointer"
-                fontSize="30"
-                color="red.300"
-                onClick={() => {
-                  deleteRow(rowIndex, row.original.id);
-                }}
-              >
-                &times;
-              </Box>
-            </Box>
-          </Box>
-        ))}
+          ))}
+        </Box>
+        <br/>
+        <Button onClick={addRow} marginBottom="15px">Add Row</Button>
+        <Text mb={2}>
+          Page {pageIndex + 1} of {items[0]?.totalPage || 1}
+        </Text>
+        <ButtonGroup size="sm" isAttached variant="outline">
+          <Button
+              onClick={previousPage}
+              isDisabled={pageIndex === 0}
+          >
+            {"<"}
+          </Button>
+          <Button
+              onClick={nextPage}
+              isDisabled={pageIndex === items[0]?.totalPage - 1}
+          >
+            {">"}
+          </Button>
+        </ButtonGroup>
+        <ToastContainer position="bottom-right" theme="colored"/>
       </Box>
-      <br />
-      <Button onClick={addRow}>Add Row</Button>
-      <Text mb={2}>
-        Page {pageIndex + 1} of {items[0]?.totalPage || 1}
-      </Text>
-      <ButtonGroup size="sm" isAttached variant="outline">
-        <Button
-          onClick={previousPage}
-          isDisabled={pageIndex === 0}
-        >
-          {"<"}
-        </Button>
-        <Button
-          onClick={nextPage}
-          isDisabled={pageIndex === items[0]?.totalPage - 1}
-        >
-          {">"}
-        </Button>
-      </ButtonGroup>
-    </Box>
   );
 };
